api/requirements.txt
requests == 2.32.5
aiohttp[speedups] == 3.13.3
psycopg == 3.3.2
psycopg_pool == 3.3.0
fastapi[standard-no-fastapi-cloud-cli] == 0.132.0
APScheduler == 3.11.2
polyline == 2.0.4
polycircles == 0.3.7
PyJWT == 2.11.0
pwdlib[argon2] == 0.3.0